Refactored Code: - created Transformer.scala class with removed dependeny on file type - added IntermediateYaml & End2EndYaml file types
