+++ sope-etl/src/main/scala/com/sope/etl/transform/YamlRunner.scala
@@ -1,6 +1,6 @@
 package com.sope.etl.transform
 
-import com.sope.etl.transform.model.YamlFile
+import com.sope.etl.transform.model.YamlFile.End2EndYaml
 import com.sope.utils.Logging
 import org.apache.commons.cli.{BasicParser, Options}
 import org.apache.spark.SparkConf
@@ -20,14 +20,16 @@ object YamlRunner extends Logging {
     val cmdLine = new BasicParser().parse(options, args, true)
     val optionMap = cmdLine.getOptions.map(option => option.getOpt.trim -> option.getValue.trim).toMap
     val mainYamlFile =  optionMap(MainYamlFileOpt)
-    val yamlTransformer = new YamlDataTransform(YamlFile(mainYamlFile))
-    logInfo("Successfully parsed YAML File, executing the Flow")
+    logInfo(s"Parsing Main Yaml File: $mainYamlFile")
+    val end2endYaml = End2EndYaml(mainYamlFile)
+    logInfo("Successfully parsed YAML File")
+    logInfo("Initializing Spark context & executing the flow..")
     val sparkConf = new SparkConf().setAppName("Spark: YAML Transformer")
     val session = SparkSession.builder()
       .config(sparkConf)
       .enableHiveSupport()
       .getOrCreate()
-    yamlTransformer.performTransformations(session.sqlContext)
+    end2endYaml.performTransformations(session.sqlContext)
   }
 
 }
